Update middleware to support any OAuth 2.0 provider
The implementation is actually not dependent on Adgangsplatformen so
lets rely on something more generic.

We could in theory use one middleware provided by an OAuth 2.0 server
but they seem to depend on the service storing users itself. It also
relies on shared public keys which we do not have at the moment.

Consequently it is worth it keeping this around.

Example
https://github.com/thephpleague/oauth2-server/blob/master/src/Middleware/ResourceServerMiddleware.php