Validate requests and return errors when invalid
By allowing the middleware to return an error we also support
descriptive error responses as we can tell the caller why the request
failed whether is was due to a missing header or no resource owner.
This was not possible when the middleware just did not add the
resource owner.

Consequently the class is renamed to validator. The logic is similar
to https://github.com/thephpleague/oauth2-server/blob/master/src/AuthorizationValidators/BearerTokenValidator.php

However this also means that the middleware must only be used for
routes which should actually be protected.

Update test cases and documentation accordingly.