Use Illuminate drivers for OAuth provider selection and configuration
Fall back to application environment for selection when not defined.
This will automatically use the most sane version for testing and
production.

We do currently only support environment variables - not Illuminate
app configuration. We cannot define $app[‘config’] for mock 
applications and Illuminate does not provide any methods for 
accessing the configuration.

This stuff belongs in environment variables anyway.