Handle users without GUIDs
Although we expect all Adgangsplatformen users to have GUIDs
experience has shown this is not always the case. When retrieving
the resource owner for some tokens users without GUIDs are returned.

In the current situation this leads to serverside errors when
services try to retrieve this id as uniqueId is not defined within
the set of attributes in the response from Adgangsplatformen.

To handle this we take several steps:

1. We mark the type of the user id as nullable for correctness
2. We allow users of the middleware to allow or deny users without
ids. The default is true to mimic the current behavior where users
without ids will also cause a failure.
3. For middleware which deny users without ids we return an error
code and message indicating the problem.

Add test case and update documentation to reflect this change.

The reason we do not automatically break for all requests without
user ids is that some token may not in fact correspond to a user.
Some services may wish to allow these nonetheless.